backlight: Add backlight handle to display to sync blanking
When the display is blanked, the backlight should be switched off two.
So I added the missing backlight = property to the display.

This fixes also the '-@ problem', where the backlight was not switched
on when the devicetree was compiled wiht symbols. The reason was simple:
the driver assumes that when the bl node has a phandle, there will be
someone holding this handle to enable and disable the backlight. The
-@/--symbols option adds a phandle to every not for alter usage, so the
drivers guess is wrong.

BCS 746-000121

(cherry picked from commit af6cbb0fabaced852373c54a0ff7daaa25d20b20)