tanaro:sn65dsi84: Use mainline sn65dsi83 driver
There is a draft for the Texas Instruments SN65DSI83/84 MIPI-to-LVDS
chips in the recent 5.14 kernel. Though it is currently an rc kernel it
can be used as a source for cherry picking.

The driver implements the SN65DSI83 chip as a DRM Bridge and thus can
use the connected panel description to configure the converter
accordingly.

This uses the new driver for the fg0700w0dsswagl1 and qp_070wsvgamlli1d
TANARO setup.  Following settings are needed:

```
+CONFIG_DRM_PANEL_LVDS=y
-CONFIG_DRM_TI_SN65DSI84=y
+CONFIG_DRM_TI_SN65DSI83=y
```

TODO:
* Review this with K&K to get a better harmonization.

BCS 746-000279

(cherry picked from commit 7c859a24f6534a8f16ca0cf74ad2474201a54c87)